Python/requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
decorator==4.4.2
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
moviepy==1.0.3
numpy==1.26.4
pillow==10.4.0
proglog==0.1.10
pydub==0.25.1
requests==2.32.3
SpeechRecognition==3.10.4
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
